requirements.txt
tiktoken
openai
python-dotenv
streamlit
numpy
pandas